requirements.txt
# Core
streamlit>=1.32
pandas>=2.2
plotly>=5.19
python-dotenv>=1.0

# Supabase (optional)
supabase

# AI features
openai
beautifulsoup4
requests
duckduckgo-search
